Fixed the bug of sensitive post still get posted by Add banned keyword validation to post and topic APIs
Moved banned keyword checks from post creation logic to API endpoints for editing posts, creating topics, and replying. This ensures that posts and topics are validated for sensitive or banned keywords at the API level, improving consistency and maintainability.

the post.js is responsible for checking when user is edting the post
the topics.js is responsible for new topics and repies
